Model class.
