+++ GamesLabCW/GamesLabCW.vcxproj
@@ -24,6 +24,7 @@
     <ClCompile Include="GameEngine.cpp" />
     <ClCompile Include="Input.cpp" />
     <ClCompile Include="main.cpp" />
+    <ClCompile Include="renderer\Model.cpp" />
     <ClCompile Include="Prototypes.cpp" />
     <ClCompile Include="renderer\Renderer.cpp" />
     <ClCompile Include="renderer\Shader.cpp" />
@@ -47,6 +48,7 @@
     <ClInclude Include="Events.h" />
     <ClInclude Include="GameEngine.h" />
     <ClInclude Include="Input.h" />
+    <ClInclude Include="renderer\Model.h" />
     <ClInclude Include="Prototypes.h" />
     <ClInclude Include="renderer\Renderer.h" />
     <ClInclude Include="renderer\Shader.h" />
