+++ GamesLabCW/GamesLabCW.vcxproj.filters
@@ -60,6 +60,9 @@
     <ClCompile Include="renderer\VBO.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="renderer\Model.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
@@ -109,6 +112,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="renderer\VBO.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="renderer\Model.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
